ai_image_detector/requirements.txt
# Core Dependencies
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Deep Learning
torch==2.1.2
torchvision==0.16.2
efficientnet-pytorch==0.7.1

# Image Processing
opencv-python==4.9.0.80
Pillow==10.2.0
scikit-image==0.22.0

# Scientific Computing
numpy==1.26.3
scipy==1.11.4

# Metadata
piexif==1.1.3

# Utilities
tqdm==4.66.1
matplotlib==3.8.2

# Data Handling
kaggle==1.5.16
